migrate-1.20-1.21 optimisation

+++ helper-scripts/migrate/migrate-1.20-1.21.py
@@ -27,9 +27,11 @@ for datasource in ("4chan", "8kun", "8chan"):
 
 	thread_constraints = [(row["column_name"], row["constraint_name"]) for row in db.fetchall("SELECT c.column_name, c.constraint_name FROM information_schema.key_column_usage AS c LEFT JOIN information_schema.table_constraints AS t ON t.constraint_name = c.constraint_name WHERE t.table_name = 'threads_%s' AND t.constraint_type = 'PRIMARY KEY';" % datasource)]
 
+	make_primary_threads_key = False
+
 	if not thread_constraints:
 		print("  No unique key detected for the threads table, adding a unique key constraint to the `id_seq` column.")
-		db.execute("ALTER TABLE threads_%s ADD PRIMARY KEY (id_seq)" % datasource)
+		make_primary_threads_key = True
 	else:
 		constraint_col, constraint_name = thread_constraints[0]
 		
@@ -37,22 +39,27 @@ for datasource in ("4chan", "8kun", "8chan"):
 			print("  id_seq already the primary key for threads %s" % datasource)
 		else:
 			print(" changing primary threads key from %s to id_seq (this might take a while)." % constraint_col)
+			make_primary_threads_key = True
 			db.execute("ALTER TABLE threads_%s DROP CONSTRAINT %s" % (datasource, constraint_name))
-			try:
-				db.execute("ALTER TABLE threads_%s ADD PRIMARY KEY (id_seq)" % datasource)
-			except UniqueViolation:
-				print("  Encountered duplicate id_seq value, rebuilding the column.")
-				db.execute("ALTER TABLE threads_%s DROP COLUMN id_seq" % datasource)
-				db.execute("ALTER TABLE threads_%s ADD COLUMN id_seq SERIAL PRIMARY KEY" % datasource)
+
+	if make_primary_threads_key:
+		try:
+			db.execute("ALTER TABLE threads_%s ADD PRIMARY KEY (id_seq)" % datasource)
+		except UniqueViolation:
+			print("  Encountered duplicate id_seq value, rebuilding the column.")
+			db.execute("ALTER TABLE threads_%s DROP COLUMN id_seq" % datasource)
+			db.execute("ALTER TABLE threads_%s ADD COLUMN id_seq SERIAL PRIMARY KEY" % datasource)
 
 	# Remove and add unique key constraints for posts table
 	print("  Checking if id_seq is the primary posts key... ")
 
 	post_constraints = [(row["column_name"], row["constraint_name"]) for row in db.fetchall("SELECT c.column_name, c.constraint_name FROM information_schema.key_column_usage AS c LEFT JOIN information_schema.table_constraints AS t ON t.constraint_name = c.constraint_name WHERE t.table_name = 'posts_%s' AND t.constraint_type = 'PRIMARY KEY';" % datasource)]
 
+	make_primary_posts_key = False
+
 	if not post_constraints:
 		print("  No unique key detected for the posts table, adding a unique key constraint to the `id_seq` column.")
-		db.execute("ALTER TABLE posts_%s ADD PRIMARY KEY (id_seq)" % datasource)
+		make_primary_posts_key = True
 	else:
 		constraint_col, constraint_name = post_constraints[0]
 		
@@ -60,13 +67,16 @@ for datasource in ("4chan", "8kun", "8chan"):
 			print("  id_seq already the primary key for posts %s" % datasource)
 		else:
 			print(" changing primary posts key from %s to id_seq (this might take a while)." % constraint_col)
+			make_primary_posts_key = True
 			db.execute("ALTER TABLE posts_%s DROP CONSTRAINT %s" % (datasource, constraint_name))
-			try:
-				db.execute("ALTER TABLE posts_%s ADD PRIMARY KEY (id_seq)" % datasource)
-			except UniqueViolation:
-				print("  Encountered duplicate id_seq value, rebuilding the entire column.")
-				db.execute("ALTER TABLE posts_%s DROP COLUMN id_seq" % datasource)
-				db.execute("ALTER TABLE posts_%s ADD COLUMN id_seq SERIAL PRIMARY KEY" % datasource)
+
+	if make_primary_posts_key:
+		try:
+			db.execute("ALTER TABLE posts_%s ADD PRIMARY KEY (id_seq)" % datasource)
+		except UniqueViolation:
+			print("  Encountered duplicate id_seq value, rebuilding the entire column.")
+			db.execute("ALTER TABLE posts_%s DROP COLUMN id_seq" % datasource)
+			db.execute("ALTER TABLE posts_%s ADD COLUMN id_seq SERIAL PRIMARY KEY" % datasource)
 
 	print("  Making indexes for post-board pairs")
